# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "V1.0.0",
    "title": "Auto Tube Guru API",
    "description": "I need to get the response from Youtube and Submit to my database, so all the calls come from my server, not the client."
  },
  "host": "localhost:3001",
  "basePath": "/listings",
  "schemes": ["http"],
  "paths": {
    "/getapikey": {
      "get": {
        "summary": "Get the next available API key",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },    
    "/getalllistings": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/getonelisting/{id}": {
      "get": {
        "description": "Get one listing",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/addlisting": {
      "post": {
        "summary": "Add a listing",
        "parameters": [
          {
            "name": "address",
            "in": "formData",
            "description": "Address of the listing",
            "required": true,
            "type": "string"
          },
          {
            "name": "beds",
            "in": "formData",
            "description": "Number of bedrooms",
            "required": true,
            "type": "integer"
          },
          {
            "name": "baths",
            "in": "formData",
            "description": "Number of bathrooms",
            "required": true,
            "type": "integer"
          },
          {
            "name": "price",
            "in": "formData",
            "description": "Price of the listing",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/editlisting/{id}": {
      "put": {
        "summary": "Update a listing",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "address",
            "in": "formData",
            "description": "Address of the listing",
            "required": true,
            "type": "string"
          },
          {
            "name": "beds",
            "in": "formData",
            "description": "Number of bedrooms",
            "required": true,
            "type": "integer"
          },
          {
            "name": "baths",
            "in": "formData",
            "description": "Number of bathrooms",
            "required": true,
            "type": "integer"
          },
          {
            "name": "price",
            "in": "formData",
            "description": "Price of the listing",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/updateapikey": {
      "put": {
        "summary": "Update API Key",
        "parameters": [
          {
            "name": "id",
            "in": "formData",
            "description": "ID of the API key",
            "required": true,
            "type": "integer"
          },
          {
            "name": "apiKey",
            "in": "formData",
            "description": "New API key value",
            "required": false,
            "type": "string"
          },
     
          {
            "name": "userCalled",
            "in": "formData",
            "description": "Number of times the API key has been called by users",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/deletelisting/{id}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}